pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "relevantai"
version = "1.0.0"
description = "RAG-powered article intelligence for Django backend"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "RelevantAI Team"}
]
keywords = ["rag", "langchain", "pinecone", "openai", "django"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: Django",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "openai>=1.0.0",
    "pinecone-client>=3.0.0,<6.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/egi03/sheep"
Documentation = "https://github.com/egi03/sheep#readme"
Repository = "https://github.com/egi03/sheep"

[tool.setuptools.packages.find]
where = ["."]
include = ["rag*"]

[tool.setuptools.package-data]
rag = ["py.typed"]
